requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
spacy==3.7.2
pyspellchecker==0.7.2
pydantic==2.5.0
python-multipart==0.0.6